% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{copymay}
\alias{copymay}
\title{Copymay}
\usage{
copymay(x, d)
}
\arguments{
\item{x}{vector containing zeros and ones only}

\item{d}{integer indicating how many elements should be taken into account}
}
\value{
Boolean. TRUE if most d elements are 1, else FALSE
}
\description{
Checks mayority of ones in first d vector elements
}
\examples{
copymay(c(1, 0, 1), 2)  # FALSE
copymay(c(1, 0, 1), 3)  # TRUE
copymay(c(1, 0, 1, 0), 4)  # FALSE
}
