% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fc_ann}
\alias{fc_ann}
\title{Forecast autoregressive neural net}
\usage{
fc_ann(x, h)
}
\arguments{
\item{x}{Time series}

\item{h}{Scalar indicating prediction horizon}
}
\value{
Double
}
\description{
User defined forecast function
}
