% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rmnoise}
\alias{rmnoise}
\title{Remove noise}
\usage{
rmnoise(x, threshold, sign = FALSE)
}
\arguments{
\item{x}{vector}

\item{threshold}{vector length 1 or 2}

\item{logical}{Default FALSE. If TRUE, negative and positive values are converted to
+1 and -1 respectively}
}
\value{
vector
}
\description{
Sets value equal to zero if between given threshold bounderies
}
\examples{
rmnoise(-3:3, threshold=2)                   # =c(-3, -2,  0, 0, 0, 2, 3)
rmnoise(-3:3, threshold=c(-1, 2))            # =c(-3, -2, -1, 0, 0, 2, 3)
rmnoise(-3:3, threshold=c(-1, 2), sign=TRUE) # =c(-1, -1, -1, 0, 0, 1, 1)
}
