% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{splitto}
\alias{splitto}
\title{Split to equal length groups}
\usage{
splitto(totlen, grplen)
}
\arguments{
\item{totlen}{Integer specifying total length}

\item{grplen}{Integer specifying length of each group}
}
\value{
list of vectors
}
\description{
Return start & end index of each group
}
\details{
Note: splitto() is similar to split() but splitto() does works fine
(no warnings) when grplen is not a multiple of totlen
}
\examples{
splitto(10, 3)    # list("1"=1:3, "2"=4:6, "3"=7:9, "4"=10)
splitto(100, 24)  # list("1"=1:24, "2"=25:48, "3"=49:72, "4"=73:96, "5"=97:100)
splitto(10, 20)   # list("1"=1:10)
}
