% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{squareme}
\alias{squareme}
\title{Squareme}
\usage{
squareme(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Numeric vector of same length as x
}
\description{
Returns the square of x
}
\examples{
squareme(2)    # = 4
squareme(1:4)  # = c(1, 4, 9, 16)
}
