% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{summe}
\alias{summe}
\title{Summe}
\usage{
summe(...)
}
\arguments{
\item{x}{A numeric vector}

\item{y}{Another numeric vector}
}
\value{
Length-one numeric vector
}
\description{
Returns the sum of any number of numeric vectors
}
\examples{
summe(1, 3, 5)       # = 9
summe(1:4, c(6, 7))  # = 23
}
