% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vlookup}
\alias{vlookup}
\title{vlookup}
\usage{
vlookup()
}
\arguments{
\item{lookup_value}{a length-one vector, value you want to lookup}

\item{table}{dataframe specifying the lookup table}

\item{lookup_colname}{a string, column name containing lookup_value}

\item{return_colname}{a string, column name containing return value}

\item{range_lookup}{a string indicating lookup method,
either "floor", "interpolate" or "ceiling"}
}
\value{
a length-one vector, specifying the value found in column labeled
with given lookup_colname
}
\description{
Lookup function similar to Excel but with two extra features: lookup column
can be any column (not just the 1st column) and includes interpolation
lookup
}
\examples{


}
