% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wf}
\alias{wf}
\title{Walk foreward analysis}
\usage{
wf(x, fcf)
}
\arguments{
\item{x}{vector or time series}

\item{fcf}{forecast funcion}

\item{window}{integer specifying rolling window}
}
\value{
vector or time series
}
\description{
Return 1 period ahead predictions using rolling training window
}
